+++ lib/cucumber/runtime.rb
@@ -234,13 +234,14 @@ module Cucumber
         # TODO: can we just use Glue::RegistryAndMore's step definitions directly?
         step_match_search = StepMatchSearch.new(@support_code.registry.method(:step_matches), @configuration)
         filters << Filters::ActivateSteps.new(step_match_search, @configuration)
+        filters << Filters::BroadcastTestCaseReadyEvent.new(@configuration)
         @configuration.filters.each { |filter| filters << filter }
         unless configuration.dry_run?
           filters << Filters::ApplyAfterStepHooks.new(@support_code)
           filters << Filters::ApplyBeforeHooks.new(@support_code)
           filters << Filters::ApplyAfterHooks.new(@support_code)
           filters << Filters::ApplyAroundHooks.new(@support_code)
-          filters << Filters::BroadcastTestCaseReadyEvent.new(@configuration)
+          # filters << Filters::BroadcastTestCaseReadyEvent.new(@configuration)
           filters << Filters::BroadcastTestRunStartedEvent.new(@configuration)
           filters << Filters::Quit.new
           filters << Filters::Retry.new(@configuration)