+++ CHANGELOG.md
@@ -18,8 +18,8 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Changed
 
-- The json formatter now reports empty scenarios. 
-  No status is reported for empty scenarios in the resulting json.
+- The JSON formatter now reports empty scenarios. 
+  No status is reported for empty scenarios in the resulting JSON.
   No more empty background is reported with empty scenarios.
   ([1533](https://github.com/cucumber/cucumber-ruby/pull/1533)
    [1530](https://github.com/cucumber/cucumber-ruby/issues/1530)