+++ CHANGELOG.md
@@ -20,6 +20,11 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
   ([1538](https://github.com/cucumber/cucumber-ruby/pull/1538)
    [1532](https://github.com/cucumber/cucumber-ruby/issues/1532)
    [aurelien-reeves](https://github.com/aurelien-reeves))
+- '--dry-run' now supports 'message' based-formatters
+  ([1537](https://github.com/cucumber/cucumber-ruby/pull/1537)
+   [1496](https://github.com/cucumber/cucumber-ruby/issues/1496)
+   [1488](https://github.com/cucumber/cucumber-ruby/issues/1488)
+   [aurelien-reeves](https://github.com/aurelien-reeves))
 - `attach` can now handle null bytes in the data.
   ([1536](https://github.com/cucumber/cucumber-ruby/pull/1536)
    [1529](https://github.com/cucumber/cucumber-ruby/issues/1529)